Gracefully handle unsupported configuration (#260)
## Motivation

As part of the 1.0 roadmap (#206), we've providing config-driven APIs,
and ones that support environment overrides. This has been in place for
a while, but there's some specific behaviour we want for unsupported
configuration, which is driven by the OTel spec. Specifically:

- When we detect unrecognised configuration in the environment
variables, we should warn, using the diagnostic logger, and not
override.
- When we detect recognised, but unsupported configuration in the
environment variables (i.e. in the OTel spec, but not supported by Swift
OTel), we'd like to have this fail differently.
- Unrecognised in-code configuration is vacuously unreachable, but for
recognised, but unsupported configuration can be marked as unilaterally
unavailable.

## Modifications

This PR is a bit on the chunky side, because configuration is naturally
a wide-spread thing, but much of it is moving existing functionality
from one place to another.

- Create an internal `protocol OTelEnvironmentRepresentable`, for types
that we want to be able to override from environment variables.
- Provide a centralised `override(...)` function that operates on types
that conform to `OTelEnvironmentRepresentable` which allows for a
consistent user experience for how successful and unsuccessful overrides
are handled and logged, including with a hint of supported values.
- Move the logic that parses types (boolean, integer, durations,
timeouts, enums, etc.) from environment variables according to the OTel
spec from `OTel+Configuration+EnvironmentOverrides.swift` to be
alongside `OTel+Configuration+EnvironmentVaraibleRepresentable.swift`.
- Add a `case none` to internal backing enums, which allows them to use
the centralised override logic with the `"none"` environment variable,
as the OTel spec requires for enum types.
- Support setting the propagators, which was hard-coded until now
(including disabling them).
- Validation for the sampling probability parameter passed to
`.traceIDRatio` (resulting in an API change).
- Extend tests to cover more invalid scenarios.

## Result

- Attempting to use configuration from the spec that is not supported
from Swift OTel will be a compile time error:

```swift
var config = OTel.Configuration.default
config.propagators = [.jaeger]  // error: 'jaeger' is unavailable: This option is not supported by Swift OTel
```

- Every successful configuration override from environment variable is
logged at info level to the diagnostic logger, with the environment
variable name, previous value, and new value—for example:
- Every failed configuration override from environment variable is
logged at warning level to the diagnostic logger, with the environment
variable name, previous value, and new value, and a hint explaining why
the value was not supported.

Here are some examples, generated by running the tests with a diagnostic
logger:

```
􀟈 Test testBatchLogRecordProcessorExportTimeout() started.
2025-07-21T12:18:51+0100 info  : environment_key=OTEL_BLRP_EXPORT_TIMEOUT environment_value=15000 new_value=15000 previous_value=30000 [OTelCore] Configuration updated from environment
2025-07-21T12:18:51+0100 info  : environment_key=OTEL_BLRP_EXPORT_TIMEOUT environment_value=0 new_value=0 previous_value=30000 [OTelCore] Configuration updated from environment
2025-07-21T12:18:51+0100 warning  : environment_key=OTEL_BLRP_EXPORT_TIMEOUT environment_value=-5000 hint=Value must be a non-negative integer number of milliseconds new_value=30000 previous_value=30000 [OTelCore] Failed to update configuration from environment
2025-07-21T12:18:51+0100 warning  : environment_key=OTEL_BLRP_EXPORT_TIMEOUT environment_value=invalid hint=Value must be a non-negative integer number of milliseconds new_value=30000 previous_value=30000 [OTelCore] Failed to update configuration from environment
􁁛 Test testBatchLogRecordProcessorExportTimeout() passed after 0.001 seconds.

...

􀟈 Test testPropagators() started.
2025-07-21T12:20:17+0100 info  : environment_key=OTEL_PROPAGATORS environment_value=b3,xray new_value=b3,xray previous_value=tracecontext [OTelCore] Configuration updated from environment
2025-07-21T12:20:17+0100 warning  : environment_key=OTEL_PROPAGATORS environment_value=b3,xray,invalid hint=Value must be a comma-separated list of values where, for each value: Value must be one of: ["tracecontext", "baggage", "b3", "b3multi", "jaeger", "xray", "otTrace", "none"] new_value=tracecontext previous_value=tracecontext [OTelCore] Failed to update configuration from environment
2025-07-21T12:20:17+0100 info  : environment_key=OTEL_PROPAGATORS environment_value=none new_value=none previous_value=tracecontext [OTelCore] Configuration updated from environment
􁁛 Test testPropagators() passed after 0.001 seconds.
```