Use matrix workflow to CI all example packages (#262)
## Motivation

We currently have two example Swift packages in the repo and plan to
rework them and add more. There are current draw backs with the testing
strategy:

1. Only one of the example projects is built in CI.
2. The example projects depend on this package using
`.package(name:path:)` using a local dependency which makes them
unsuitable for lifting out of this repo.

## Modifications

- ci: Create matrix workflow for examples
- examples: Use upstream URL in example projects
- ci: Remove now vestigial, non-matrix job for counter example
- examples: Get HTTP server example passing CI


## Result

- CI now uses a matrix job to auto-discover, build, and test all example
packages in `Examples/`.
- Example packages now depend on Swift OTel by URL so they're suitable
for using by others.

## Notes

This PR contains the bare minimum to get the server example building in
CI. We do want to iterate on it (and probably add more examples) but we
should decouple that from this PR to get the CI in place.