Add fan-in CI job to collect examples matrix jobs (#269)
## Motivation

In #262 we added a GitHub Workflow to discover example projects and
create a matrix job. This is great, as we don't have to configure CI for
new examples. However, it's a pain for keeping the branch protection
rules up to date. However, we can do better: after creating the matrix
job to fan-out the examples, we can have a job that depends on the
matrix, which will fan-in the results. Then we can have that collector
job be _required_ and the other ones will be implicitly required.

## Modifications

Add an `Example package (All)` pseudo job that depends on the matrix job
for the examples

## Result

We have a single, deterministically named CI job that we can add to the
branch protection rules, but still get required CI for any new examples
added to the repo.