Add missing trait compiler guards to fix macOS adopters (#264)
## Motivation

When developing the examples, for which there's an upcoming PR, I found
there were some missing trait guards, meaning that you couldn't compile
the project with lower platform. This is because it was unilaterally
trying to use the gRPC symbols in the bootstraps. The only way we can
resolve this is to introduce an existential at this layer. This is no
worse than the behaviour we had until very recently, and still a
significant improvement on having existential everywhere. Here they only
exist for the exporter (not on the hot path for adopter code), and only
when using gRPC (the price we'll have to pay to lower the minimum
platform requirements for the package).

## Modifications

- Add missing trait compiler guards

## Result

Downstream packages can compile without raising their minimum platform.

## Notes

Upcoming PRs will flesh out the examples, which will help us catch some
of this. But the real reason this wasn't caught was because there's no
macOS CI, since availability annotations are only relevant on Apple
platforms.