Update HTTP server example with logging backend (#265)
## Motivation

As part of the 1.0 roadmap (#206), we want to refresh and extend the
collection of example projects to show how to use the 1.0 API and
integrate with the wider ecosystem of packages and tools. Now that the
1.0 API is in place, we can start doing that.

This PR makes a start by updating the existing HTTP server example to
use the latest API and also to use the logging backend.

## Modifications

- Update HTTP server example with logging backend
- Rename to make space for more examples
- Update the README with improved diagram including logging export

## Result

The HTTP server example now compiles and uses the latest API for logs,
metrics, and traces.