Refresh README with overview and getting started (#266)
## Modifications

As we approach 1.0 (see #206), we need to update the documentation to
reflect the new package structure and APIs. That includes the README.

## Modifications

- Refresh README with overview and getting started

## Result

README now accurate for 1.0 alpha.

## Notes

- The new structure follows the latest style being rolled out in some of
the SSWG repos, including Swift Log and gRPC Swift 2 (relevant to this
project).
- The SPI config needs updating, which will be done in a separate PR,
and will likely require us to switch the default branch back to `main`.
- This also makes a headline callout to what this package is: _OTLP
backend for Swift Log, Swift Metrics, and Swift Distributed Tracing._