Add covering README to Examples/ directory (#270)
## Motivation

As we approach 1.0 (#206), we'd like to flesh out the examples
directory. As a result, it would be useful if this directory, itself,
had some structure and content.

## Modifications 

Add covering README to Examples/ directory with some TODOs that we can
fill in.

## Result

People landing on in the examples directory (e.g. from docs) will have
some guidance.