Rework DocC catalog for new package structure and update SPI config (#271)
## Motivation

We're approaching the 1.0 alpha, and since the 0.x branch we have
significantly changed the API surface and package structure (see #205).
We now need to get our DocC catalog to reflect this.

We need to play some tricks to have the DocC catalog play nicely with
the new Swift OTel package structure.

The only product module, `OTel`, doesn't actually provide any public
top-level symbols. Rather it provides public
extensions on `enum OTel`, which actually lives in the internal
`OTelCore` module, and makes this available using
`@_exported import OTelCore.OTel`. The end experience is that folks just
`import OTel` and they get the API.

By default, this leads to a sub-par DocC experience, but DocC exposes
enough levers that we can workaround it and
present the reader with the same illusion: that there's one
module/symbol, `OTel`, where all the APIs live.

We do this by using a careful combination `@Options` and `@Metadata`,
which have been documented in code comments.

## Modifications

- Rework DocC bundle for new module, with reduced API
- Add CI job for checking DocC documentation

## Result

- DocC documentation now builds, without errors, for the product module.
- DocC documentation now presents the API as we expect users to use it.
- CI now checks that DocC documentation continues to build.