Add missing shared swift settings to OTelTests target (#274)
## Motivation

When doing a quick audit of the package manifest, I noticed that the
`OTelTests` test target is not using the `sharedSwiftSettings` used by
all other targets. This means it won't be able to use the
`@available(gRPCSwift, *)` macro and any other compile time settings we
wish to enforce across the whole package.

## Modifications

- Add missing shared swift settings to OTelTests target

## Result

The `OTelTests` target is compiled with the same settings as all other
targets.