Fix default endpoint when using OTLP/gRPC exporter (#276)
## Motivation

The default endpoint in the configuration datamodel is the
`http://localhost:4318`, as defined by the OTel spec. This is the
default endpoint for the OTLP/HTTP exporter. When using OTLP/gRPC and
not explicitly setting an endpoint, the spec says that
`http://localhost:4317` should be used.

## Modifications

- Add test for default OLTP/gRPC endpoint (both in-code, and environment
overrides)
- Fix default endpoint for OTLP/gRPC endpoint.

## Result

When using OTLP/gRPC the endpoint will default to
`http://localhost:4317`.