Add HTTP server examples using OTLP/HTTP+json and OTLP/gRPC (#277)
## Motivation

In #265 we updated the HTTP server example and in #270 we added a README
to the Examples directory with some placeholders. The next step was to
iterate through the list and flesh out the examples.

This PR adds an equivalent HTTP server example using OTLP/HTTP+json and
OTLP/gRPC.

## Modifications 

- Rename example: hello-world-hummingbird-server ->
hello-world-hummingbird-server-otlp-http-protobuf
- Fix pre-canned links in hello-world example
- Add example: hello-world-hummingbird-server-otlp-http-json
- Add example: hello-world-hummingbird-server-otlp-grpc

## Result

- Jaeger and Prometheus links in example README use correct query
- Examples now include OTLP/HTTP+json and OTLP/gRPC

## Notes

This PR depends on #275 and #276.