Throw error if (m)TLS config contains files that do not exist (#279)
## Motivation

When using the OLTP exporters with TLS or mTLS, the user must pass paths
to files. If these files do not exist, we currently leave it up to the
underlying library to fail gracefully (or not...). We can validate these
as part of the client creation and throw a curated error instead.

## Modifications

Throw error if (m)TLS config contains files that do not exist

## Result

Paths are validated at construction.