Remove Counter example (#283)
## Motivation

We've curated a list of examples for the 1.0 alpha release, but the old
counter example was still hanging out there. We may choose to bring a
server-less example back in the future, but for now we can remove it.

## Modifications

Remove Counter example

## Result

Only the recent, curated examples are present for alpha.