Enable link to docs in README now SPI has been updated (#284)
## Motivation

The link to the docs in the README and the DocC landing page was
temporarily disabled until we updated the Swift Package Index config.
That was updated already in #271 so we can enable it again.

## Modifications

- Enable link to docs in README now SPI has been updated

## Result

Link to docs in README and DocC landing page now works.

## Notes

It takes you to the latest release on SPI, which will be the alpha soon
(it was only just tagged so SPI is a little behind).