Fix tag used in example projects (#285)
## Motivation

I cannot type and so in #282 I misspelled the tag in the example
projects. Our fan-in on the examples CI didn't check the results so
auto-merge didn't catch it.

## Modifications

Fix the tag used.

## Result

Examples should build in CI.

## Notes

I will not use auto-merge for this PR and will work on making the fan-in
on the CI work better today.