Update docs link in README and DocC root to point to main branch docs (#286)
## Motivation

We're currently using
https://swiftpackageindex.com/swift-otel/swift-otel/documentation/otel
in our README and DocC root, which redirects to the "latest release".
Unfortunately, SPI considers this to be 0.12.0 and not 1.0.0-alpha.1,
which it considers the "latest beta release".

## Modifications

Update docs link in README and DocC root to point to main branch docs

## Result

The link from the README and the DocC root will point to more relevant
docs.