Check results of matrix jobs in Example Package (All) CI job (#287)
## Motivation

The CI has an `Example Package (All)` job which depends on the matrix
job for checking the example projects. It wasn't actually checking the
result of the matrix job though, so it would be skipped, even if one or
more example matrix jobs failed.

## Modifications

- Check the results of the matrix job in the fan-in CI job

## Result

CI will gate against failed examples.

## Notes

This PR will contain one commit that deliberately poisons the examples,
so we can see the issue (the All job will not be marked as failed). Then
it will be updated with a fix for the All job (which will then fail, as
required). Then we'll revert the commit that poisons the examples.