Add support for enabled backends with no exporter (#291)
## Motivation

The OTel spec defines support for a backend that is enabled but without
an exporter, and even supports the environment variable overrides for
this, e.g. `OTEL_TRACES_EXPORTER=none`. Until now, this has been
conflated with disabling the backend altogether.

However, this is not desirable since we have a specific use case that
requires exactly this. If a user wants to have logs with span and trace
IDs but is not interested in exporting traces (regardless of whether
they use the OTLP logging backend or just the logging metadata provider
with their own backend).

We even have an end-to-end test for this use case already, added in
#256, but this test bootstraps the tracing backend with the default
configuration. While that will be sufficient, it will result in an OTLP
exporter continually running and not being able to export.

We should instead support having the backends enabled but not exporting.

Thankfully, because we have the generic wrapper types, which we use from
the public API, this is pretty trivial, and also removes a precondition
failure for code that we claimed was previously unreachable.

We also already had defined `case none` in the backing enum in the
config API (so we can detect it from environment variables), but it
wasn't reachable by the public configuration API.

## Modifications

- Support `case none` in the wrapped exporter enums.
- Add public API for
`Configuration.[Logs,Metrics,Traces].ExporterSelection.none`
- Stop conflating `OTEL_<SIGNAL>_EXPORTER=none` with disabling the
backend entirely
- Update the configuration tests
- Update the end-to-end that uses the logging metadata provider to
correlate logs with active spans to not require an exporter.

## Result

- Can now bootstrap backends without an exporter
- New API `Configuration.Logs.ExporterSelection.none`
- New API `Configuration.Metrics.ExporterSelection.none`
- New API `Configuration.Traces.ExporterSelection.none`
- Support for `OTEL_LOGS_EXPORTER=none`
- Support for `OTEL_METRICS_EXPORTER=none`
- Support for `OTEL_TRACES_EXPORTER=none`