Add support for console logs exporter (#292)
## Motivation

The OTel spec defines a console exporter for all three signals, which
are intended to be used for local development[^1]. This is particularly
useful for logs because, once the Swift logging subsystem is
bootstrapped with an OTLP logger, no further logging will be visible on
the console so it would be good for adopters to be able to choose a
console logs exporter, either using the configuration API, or by
`OTEL_LOGS_EXPORTER=console`.

The spec suggests a few requirements for the console exporter:

1. It should use the simple processor (not the batch processor)
2. It should print a warning when being used (and provides the wording)

We already had the simple processor implementation in the `OTelTesting`
module, but it looks like it will need a little work, since it does not
currently call its exporter's run method.

We also already had the configuration in place for this, but was marked
always unavailable, since we didn't support it, so adding the new
configuration API is trivial.

## Modifications

- Move `OTelSimpleLogRecordProcessor` from `OTelTesting` to `OTelCore`
- Add diagnostic logging to `OTelSimpleLogRecordProcessor`
- Update `SimpleLogRecordProcessor` to actually run its exporter
- Add `OTelConsoleLogRecordExporter` that just prints to stdout
- Format the log messages in the console exporter
- Add `console` case to `WrappedLogRecordExporter` enum
- Add `WrappedLogRecordProcessor` with automatic selection based on
exporter
- Add tests for automatic processor selection based on exporter
- Enable console logs exporter in the configuration API
- Add end-to-end test for console logging exporter

## Result

- New API:
`OTel.Configuration.LogsConfiguration.ExporterSelection.console`
- Support for `OTEL_LOGS_EXPORTER=console`

[^1]:
https://opentelemetry.io/docs/specs/otel/logs/sdk_exporters/stdout/