Restructure package to avoid module leaks, fix traits, and fix docs (#294)
## Motivation

Apologies in advance, this'll be a big patch, and an even bigger
writeup...

Our goal for the 1.0 PR was to have people simply `import OTel` and to
then have the APIs there, e.g. `OTel.bootstrap`.

In order to preserve the historic structure of the package, we acheived
this by making all previous targets internal and providing a new `OTel`
product/target which can expose various APIs based on traits (e.g.
OTLP/HTTP and OTLP/gRPC support). However, there were types that needed
to be used by both the public and internal modules, so to break the
circular dependency, these were placed in the, now internal, `OTelCore`
module and we used an `@_exported import OTelCore.OTel` in the `OTel`
module.

I concede that all this was a poor idea, and it is showing its
limitations in at least the following ways:

1. The `OTelCore` module is leaking (it's there whether we like it or
not).
2. DocC cannot cope with this package gymnastics and there's simply no
way to get documentation for the configuration API, which is technically
in `OTelCore`. This can be seen on the currently available docs on SPI,
which do not have any rendered docs for `Configuration`.
3. The package isn't even achieving its goal. Concretely, if we try to
thin the used traits in the examples, they (a) don't build for 6.1
(oops!) and (b) don't reduce the dependency graph for 6.2 (double
oops!).

We need to revert the package structure to something more conventional.

## Modifications

First, given our desire for a single module that uses traits to impact
API visibility/availability without leaking the existence of other
modules, we squash the modules into one. We've preserved the directory
hierarchy though, this has required:

- Updating the Makefile to generate code into the shared module
- Adding a step to the Makefile to add appropriate compiler guards
around files
- Removing the modulemap, no longer needed by the protobuf plugin
- Rerunning codegen to get generated code in the shared module, with
trait guards
- Moving handwritten files into the shared module
- Removing the internal targets
- Removing any imports of those removed targets

This then gave us the ability to fix the docs, where we:

- Remove all the hacks we were applying to the DocC bundle
- Add some lightweight curation to the symbols (more can follow)

Finally, we want to make sure that (a) the example projects build on
both Swift versions and (b) that the traits work as required:

- Update the CI to build examples in both 6.1 and 6.2
- Update the protocol-specific examples to use only the required trait
- Add a script that can validate the dependencies of swift-otel used in
an example based on traits

Support for trait-based dependency resolution is _very_ new (landed 2
weeks ago in main) and it's not yet working correctly on Linux, so we'll
skip it for now, but I'll show it working in the results section.

## Results

- There is now only one module, `OTel`, in the `OTel` product
- CI now shows that examples can build with traits
- CI now builds examples with both 6.1 and 6.2
- We have a script that will trait-based dependencies when they're
working
- The docs are working (will add screenshot to PR comments)

## Notes

Here we have `Examples/hello-world-hummingbird-server-otlp-http-json`,
which contains a local dependency on swift-otel, which I've run with
various traits and then checked the dependencies, and then run the
script:

#### Traits unspecified (default traits)

```console
% grep '\.package' Package.swift | grep swift-otel
        .package(name: "swift-otel", path: "../.."),

% swift package show-dependencies --format json | jq -r '.dependencies[] | select((.identity | startswith("swift-otel"))) | .dependencies[].identity' | grep -e grpc -e http -e protobuf
swift-protobuf
grpc-swift-2
grpc-swift-nio-transport
grpc-swift-protobuf
async-http-client

% ../..** Checking traits used for swift-otel dependency...
** Example package depends on swift-otel with the following traits:
default

** Traits include default; checking for expected HTTP and gRPC dependencies...
** ✅ Found no errors./scripts/check-example-dependencies.sh
```

#### `traits: ["OTLPHTTP", "OTLPGRPC"]`

```console
% grep '\.package' Package.swift | grep swift-otel
        .package(name: "swift-otel", path: "../..", traits: ["OTLPHTTP", "OTLPGRPC"]),
% swift package show-dependencies --format json | jq -r '.dependencies[] | select((.identity | startswith("swift-otel"))) | .dependencies[].identity' | grep -e grpc -e http -e protobuf
swift-protobuf
grpc-swift-2
grpc-swift-nio-transport
grpc-swift-protobuf
async-http-client

% ../../scripts/check-example-dependencies.sh
** Checking traits used for swift-otel dependency...
** Example package depends on swift-otel with the following traits:
default

** Traits include default; checking for expected HTTP and gRPC dependencies...
** ✅ Found no errors.
```

#### `traits: ["OTLPHTTP"]`

```console
% grep '\.package' Package.swift | grep swift-otel
        .package(name: "swift-otel", path: "../..", traits: ["OTLPHTTP"]),

% swift package show-dependencies --format json | jq -r '.dependencies[] | select((.identity | startswith("swift-otel"))) | .dependencies[].identity' | grep -e grpc -e http -e protobuf
swift-protobuf
async-http-client

% ../../scripts/check-example-dependencies.sh
** Checking traits used for swift-otel dependency...
** Example package depends on swift-otel with the following traits:
OTLPHTTP

** Traits include OTLPHTTP; checking for expected HTTP dependencies...
** Traits does NOT contain OTLPGRPC or default; checking for unexpected gPRC dependencies...
** ✅ Found no errors.
```

#### `traits: ["OTLPGRPC"]`

```console
% grep '\.package' Package.swift | grep swift-otel
        .package(name: "swift-otel", path: "../..", traits: ["OTLPGRPC"]),

% swift package show-dependencies --format json | jq -r '.dependencies[] | select((.identity | startswith("swift-otel"))) | .dependencies[].identity' | grep -e grpc -e http -e protobuf
swift-protobuf
grpc-swift-2
grpc-swift-nio-transport
grpc-swift-protobuf

% ../../scripts/check-example-dependencies.sh
** Checking traits used for swift-otel dependency...
** Example package depends on swift-otel with the following traits:
OTLPGRPC

** Traits include OTLPGRPC; checking for expected gRPC dependencies...
** Traits does NOT contain OTLPHTTP or default; checking for unexpected HTTP dependencies...
** ✅ Found no errors.
```

#### `traits: []`

```console
% grep '\.package' Package.swift | grep swift-otel
        .package(name: "swift-otel", path: "../..", traits: []),

% swift package show-dependencies --format json | jq -r '.dependencies[] | select((.identity | startswith("swift-otel"))) | .dependencies[].identity' | grep -e grpc -e http -e protobuf

% ../../scripts/check-example-dependencies.sh
** Checking traits used for swift-otel dependency...
** Example package depends on swift-otel with the following traits:


** Traits does NOT contain OTLPHTTP or default; checking for unexpected HTTP dependencies...
** Traits does NOT contain OTLPGRPC or default; checking for unexpected gPRC dependencies...
** Traits does NOT contain OTLPHTTP, OTLPGRPC, or default; checking for unexpected HTTP or gRPC dependencies...
** ✅ Found no errors.
```