Use canary service to support tests of service lifecycle graceful shutdown (#299)
## Modifications

Right now Swift Service Lifecycle does not have API to wait for a
service group to be running. This can make writing tests of the graceful
shutdown behaviour hard. For example, the following naive test will
probabilistically fail because run might try to run the the services
after the shutdown:

```swift
@Test func test() async throws {
    struct S: Service {
        var name: String
        func run() async throws {
            print("\(name) running")
            await AsyncStream.makeStream(of: Void.self).stream.cancelOnGracefulShutdown().first { true }
            print("\(name) done")
        }
    }

    let serviceGroup = ServiceGroup(services: [S(name: "A"), S(name: "B")], logger: Logger(label: "group"))
    try await withThrowingTaskGroup { group in
        group.addTask { try await serviceGroup.run() }
        await serviceGroup.triggerGracefulShutdown()
        try await group.waitForAll()
    }
}
```

```
Caught error: ServiceGroupError: errorCode: The service group has already finished running the services.
```

We have a bunch of very fragile substitutes for this in our tests right
now. For example, we'll hook into a test clock and wait for a specific
number of sleep calls. As I've been refactoring to make more robust
shutdown logic, these have been a problem.

This initial PR adds a different mechanism for waiting for the service
group to be running before we shut it down in our tests of graceful
shutdown.

## Modifications

- Move Logging+OTelDisabled to Logging+TestHelpers and add ._otelDebug
- Add canary service to support tests of graceful shutdown

## Result

Less fragile graceful shutdown tests.