Ensure all run methods only complete on shutdown/cancel (#301)
## Motivation

In #236 we added a run method requirement to exporter protocols to
support the OTLP/gRPC exporter migration to gRPC Swift 2, which requires
a run method for the connection pool. As such, there became a need for
the exporters to be run. This is currently done by the processor, but
there are issues with the way the composition of these run methods works
with graceful shutdown. Specifically, the main issue we have right now
is how the processor behaves when the exporter exits early (cleanly or
throws).

As part of trying to address that (future PR to come), I noticed that
most of the exporters had an empty, no-op run method. This isn't great
because anything that tries to formalise the shutdown semantics of the
processor when the exporter early-exits will trigger for these exporters
that immediately return from their run methods.

The semantics of a type with a run method should be that it "blocks"
until the "background work" is finished—in our cases when the type is
cancelled or shutdown.

## Modifications

- Ensure all run methods only complete on shutdown/cancel

## Result

All types with run method have usual semantics, which will unblock
making the error handling of exporters more robust.