Lean into ServiceGroup for consistent handling of shutdown of nested services (#304)
## Motivation

Some of the backends have a set of nested types, that each have a run
method. An example is the tracer, which holds a processor, which holds
an exporter. Currently, we rely on the outer type to run the inner
type's run method as part of the outer type's run method. However, this
has some sharp edges.

Consider the following pattern:

```swift
func run() async throws {
    withThrowingTaskGroup { group in
        group.add { try await self.inner.run() }
        for try await _ in timer {
            // outer work
            // ...
        }
        // outer work done, shut down
        // ...
        try await group.waitForAll()
    }
}
```

If the inner run method finishes early, either by throwing or by exiting
cleanly, then nothing will notice this. The outer type will just carry
on, as normal.

We _could_ fix this by a judicial use of `try await group.next()`. This
will cancel the remaining child tasks and throw the error if the child
task throws, but if it exits cleanly then you'll need some other
construction, something like this:

```swift
func run() async throws {
    withThrowingTaskGroup { group in
        group.add { try await self.inner.run() }
        for try await _ in timer {
            // outer work
        }
        // outer work done, shut down
        do {
            try await group.next()
            // shutdown logic
        } catch {
            // shutdown logic
        }
    }
}
```

At this point, we're reinventing the wheel and probably still have some
corner cases that aren't covered well.

This is the whole purpose of `ServiceGroup`—to keep track of a set of
related tasks and coordinate their lifetimes.

Since the types in each backend share a fate, we can use a nested group
to represent that service. We can also leverage the graceful shutdown
termination behaviour if one of them fails which is nice. For example,
if the exporter fails, there's no hope for any additional telemetry
being exported on graceful shutdown, but, if the processor failed
instead, then allowing the exporter to flush might give a little more
useful observability for services.

Leaning into this pattern allows for each type to pretty much eliminate
its inner task group and devote its run method only to its own work,
which makes it much easier to understand.

## Modifications

- Add tests for BLRP shutdown when exporter exits or throws
- Add tests for PEMR shutdown when exporter exits or throws
- Add tests for BSP shutdown when exporter exits or throws
- Use nested service group to manage lifecycle of logging backend
services
- Use nested service group to manage lifecycle of metrics backend
services
- Use nested service group to manage lifecycle of tracing backend
services

## Result

Much more structured shutdown behaviour of nested types.