Emit info diagnostic to console when bootstrapping logs (#306)
## Motivation

If an adopter is using Swift OTel with the OTLP logging backend then
their local development experience changes significantly because all
logging that uses the bootstrapped backend—from their own application
and all the libraries they depend on—will no longer be printed to the
console, but be exported to the OTLP endpoint. It's likely that users
will not want to set up a local OTLP endpoint just to see their logs
during development either.

We now have the console log exporter, added in #292, which is exactly
for this purpose. However, we want to make sure that adopters get a
great out-of-the box experience when they try out the package, so we
should signpost this when we bootstrap the logging system with an OTLP
backend.

## Modifications 

- Emit info diagnostic to console when bootstrapping logs

## Result

When users call `bootstrap`, a prominent message will be emitted to the
console explaining how to enable console logging for local development.