Drain incoming streams in batch processors during graceful shutdown (#310)
## Motivation

Our graceful shutdown is currently not very... graceful. We were using
`AsyncSequence.cancelOnGracefulShutdown` a little too zealously. This
was masked because until recently we didn't have very good end-to-end
shutdown tests. We recently added a canary to allow the service to have
at least started before triggering shutdown, but this masked the lack of
grace in the shutdown. Specifically, the delay it added allowed the logs
and spans to be flushed _before_ graceful shutdown was triggered, so we
couldn't tell what would happen during shutdown if they were still in
the internal streams.

This manifested when I ran the test thousands of times, eventually we'd
find in the end-to-end tests that we'd have fewer spans or log records
than expected. In these cases, the batch processor's incoming stream had
been cancelled so never made it into the buffer before force flush was
called.

Instead of cancelling these streams using `cancelOnGracefulShutdown`, we
should use a graceful shutdown handler to finish the streams cleanly. If
the shutdown escalates to a cancellation then so be it, but if the user
has chosen to trigger a graceful shutdown, we should do our best to not
lose telemetry.

## Modifications

- Use debug logger in tests now we have one that does not conflate user
logging
- Use `withGracefulShutdownHandler` instead of
`cancelOnGracefulShutdown`

## Result

On shutdown the incoming streams in the batch span and log record
processors are drained rather than dropped.